LOLWUT: split the command from version-specific implementations.
